Remove cast

+++ src/test/java/org/plumelib/javacparse/JavacParseTest.java
@@ -195,7 +195,7 @@ class JavacParseTest {
 
   @Test
   void memoryTest() {
-    long initialUsedMemory = (double) SystemPlume.usedMemory(true);
+    long initialUsedMemory = SystemPlume.usedMemory(true);
     int numIterations = 10; // Each iteration takes approximately 1 second.
     if (System.getenv("GITHUB_HEAD_REF") != null) {
       // If this line is reached, the program is running in GitHub Actions continuous integration.
